Refactor nat building into utils.
